requirementse.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
dynaconf==3.1.4
filelock==3.0.12
gunicorn==20.0.4
hvac==0.10.8
idna==2.10
knox==0.1.11
loguru==0.5.3
Markdown==3.3.4
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
pypiwin32==223
pytz==2021.1
pywin32==300
qrcode==6.1
requests==2.25.1
screen-brightness-control==0.8.5
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.2
whitenoise==5.2.0
win10toast==0.9
win32-setctime==1.0.3
WMI==1.5.1
